+++ packages/components/modal/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @nextui-org/modal
 
+## 0.0.0-dev-v2-20230601001842
+
+### Patch Changes
+
+- Button and spinner styles fixed
+- Updated dependencies
+  - @nextui-org/framer-transitions@0.0.0-dev-v2-20230601001842
+  - @nextui-org/shared-icons@0.0.0-dev-v2-20230601001842
+  - @nextui-org/shared-utils@0.0.0-dev-v2-20230601001842
+  - @nextui-org/use-aria-button@0.0.0-dev-v2-20230601001842
+  - @nextui-org/use-disclosure@0.0.0-dev-v2-20230601001842
+  - @nextui-org/dom-utils@0.0.0-dev-v2-20230601001842
+  - @nextui-org/system@0.0.0-dev-v2-20230601001842
+  - @nextui-org/theme@0.0.0-dev-v2-20230601001842
+
 ## 0.0.0-dev-v2-20230531020353
 
 ### Patch Changes