+++ packages/utilities/shared-icons/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nextui-org/shared-icons
 
+## 0.0.0-dev-v2-20230601001842
+
+### Patch Changes
+
+- Button and spinner styles fixed
+
 ## 0.0.0-dev-v2-20230531020353
 
 ### Patch Changes